Move L-BFGS-B to C_Libraries
